+++ src/parse-engines/parse-engine-registry.ts
@@ -1,15 +1,16 @@
 import IParseEngine from "./common/parse-engine";
-import Primefaces8ParseEngine from "./types/primefaces8-parse-engine";
-import Primefaces10ParseEngine from "./types/primefaces10-parse-engine";
-import HFaceletsParseEngine from "./types/h-facelets-parse-engine";
-import FFaceletsParseEngine from "./types/f-facelets-parse-engine ";
-import CFaceletsParseEngine from "./types/c-facelets-parse-engine";
-import CCFaceletsParseEngine from "./types/cc-facelets-parse-engine";
-import UIFaceletsParseEngine from "./types/ui-facelets-parse-engine";
-import RichfacesParseEngine from "./types/richfaces45-engine";
-import RichfacesA4JParseEngine from './types/richfaces45-a4j-engine';
-import OmnifacesFaceletsParseEngine from "./types/omnifaces-facelets-parse-engine";
-import PrimefacesExtensionsFaceletsParseEngine from "./types/primefacesExtensions-facelets-parse-engine";
+import Primefaces8 from "./types/primefaces8";
+import Primefaces10 from "./types/primefaces10";
+import Primefaces11 from "./types/primefaces11";
+import H from "./types/h";
+import F from "./types/f";
+import C from "./types/c";
+import CC from "./types/cc";
+import UI from "./types/ui";
+import Richfaces from "./types/richfaces45";
+import RichfacesA4J from './types/richfaces45-a4j';
+import Omnifaces from "./types/omnifaces";
+import PrimefacesExtensions from "./types/primefacesExtensions";
 
 class ParseEngineRegistry {
     public static getParseEngine(taglibId: string): IParseEngine {
@@ -31,17 +32,18 @@ class ParseEngineRegistry {
 
     private static languagesIds: string[];
     private static registry: IParseEngine[] = [
-        new Primefaces8ParseEngine(),
-        new Primefaces10ParseEngine(),
-        new RichfacesParseEngine(),
-        new RichfacesA4JParseEngine(),
-        new HFaceletsParseEngine(),
-        new FFaceletsParseEngine(),
-        new CFaceletsParseEngine(),
-        new CCFaceletsParseEngine(),
-        new UIFaceletsParseEngine(),
-        new OmnifacesFaceletsParseEngine(),
-        new PrimefacesExtensionsFaceletsParseEngine()
+        new Primefaces8(),
+        new Primefaces10(),
+        new Primefaces11(),
+        new Richfaces(),
+        new RichfacesA4J(),
+        new H(),
+        new F(),
+        new C(),
+        new CC(),
+        new UI(),
+        new Omnifaces(),
+        new PrimefacesExtensions()
     ];
 }
 export default ParseEngineRegistry;