+++ src/parse-engines/types/cc.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/richfaces45.json';
+import * as data from '../data/cc.json';
 
-class RichfacesParseEngine implements IParseEngine {
-    public taglibId = "richfaces45";
-    public parse(version: string): ComponentDefinition[] {
+class CC implements IParseEngine {
+    public taglibId = "cc";
+    public parse(version: string):ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default RichfacesParseEngine;
+export default CC;