+++ src/parse-engines/types/h.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/cc.json';
+import * as data from '../data/h.json';
 
-class CCFaceletsParseEngine implements IParseEngine {
-    public taglibId = "cc";
-    public parse(version: string):ComponentDefinition[] {
+class H implements IParseEngine {
+    public taglibId = "h";
+    public  parse(version: string): ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default CCFaceletsParseEngine;
+export default H;