+++ src/parse-engines/types/omnifaces.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/primefaces-extensions.json';
+import * as data from '../data/omnifaces.json';
 
-class PrimefacesExtensionsFaceletsParseEngine implements IParseEngine {
-    public taglibId = "primefaces-extensions";
+class Omnifaces implements IParseEngine {
+    public taglibId = "omnifaces";
     public parse(version: string): ComponentDefinition[]{
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default PrimefacesExtensionsFaceletsParseEngine;
+export default Omnifaces;