+++ src/parse-engines/types/primefaces10.ts
@@ -1,0 +1,12 @@
+import IParseEngine from "../common/parse-engine";
+import ComponentExtractor from "../common/component-extractor";
+import ComponentDefinition from "../../common/component-definition";
+import * as data from '../data/primefaces10.json';
+
+class Primefaces10 implements IParseEngine {
+    public taglibId = "primefaces10";
+    public parse(version: string): ComponentDefinition[] {
+        return ComponentExtractor.extract(JSON.stringify(data.components.component));
+    }
+}
+export default Primefaces10;