+++ src/parse-engines/types/primefaces8.ts
@@ -3,10 +3,10 @@ import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
 import * as data from '../data/primefaces8.json';
 
-class Primefaces8ParseEngine implements IParseEngine {
+class Primefaces8 implements IParseEngine {
     public taglibId = "primefaces8";
     public parse(version: string): ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default Primefaces8ParseEngine;
+export default Primefaces8;