+++ src/parse-engines/types/primefacesExtensions.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/ui.json';
+import * as data from '../data/primefaces-extensions.json';
 
-class UIFaceletsParseEngine implements IParseEngine {
-    public taglibId = "ui";
-    public parse(version: string): ComponentDefinition[] {
+class PrimefacesExtensions implements IParseEngine {
+    public taglibId = "primefaces-extensions";
+    public parse(version: string): ComponentDefinition[]{
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default UIFaceletsParseEngine;
+export default PrimefacesExtensions;