+++ src/parse-engines/types/richfaces45-a4j.ts
@@ -3,10 +3,10 @@ import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
 import * as data from '../data/richfaces45-a4j.json';
 
-class RichfacesA4JParseEngine implements IParseEngine {
+class RichfacesA4J implements IParseEngine {
     public taglibId = "richfaces45-a4j";
     public parse(version: string): ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default RichfacesA4JParseEngine;
+export default RichfacesA4J;