+++ src/parse-engines/types/richfaces45.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/omnifaces.json';
+import * as data from '../data/richfaces45.json';
 
-class OmnifacesFaceletsParseEngine implements IParseEngine {
-    public taglibId = "omnifaces";
-    public parse(version: string): ComponentDefinition[]{
+class Richfaces implements IParseEngine {
+    public taglibId = "richfaces45";
+    public parse(version: string): ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default OmnifacesFaceletsParseEngine;
+export default Richfaces;