+++ src/parse-engines/types/ui.ts
@@ -1,12 +1,12 @@
 import IParseEngine from "../common/parse-engine";
 import ComponentExtractor from "../common/component-extractor";
 import ComponentDefinition from "../../common/component-definition";
-import * as data from '../data/f.json';
+import * as data from '../data/ui.json';
 
-class FFaceletsParseEngine implements IParseEngine {
-    public taglibId = "f";
+class UI implements IParseEngine {
+    public taglibId = "ui";
     public parse(version: string): ComponentDefinition[] {
         return ComponentExtractor.extract(JSON.stringify(data.components.component));
     }
 }
-export default FFaceletsParseEngine;
+export default UI;